Moves into its own component infoMovesDisplay
