+++ src/component/info/infoDisplay.tsx
@@ -68,25 +68,10 @@ export default class InfoDisplay extends React.Component<Props> {
 
 	render() {
 		const types: string[] = [];
-		const moves: string[] = [];
-		const flavorText: string [] = [];
-		let movesIndex: number = 0;
-
 		this.props.pokemon.types?.forEach(type => {
 			types.push(type.type.name);
 		});
-		this.props.pokemon.moves?.forEach(move => {
-			moves.push(move.move.name);
-		});
 
-		this.props.pokemon.movesFlavorText?.forEach(text => {
-			flavorText.push(text);
-		});
-		
-		function createFlavorText (i: number) {
-				movesIndex ++;
-				return  flavorText[i];
-		}
 
 		
 
@@ -101,20 +86,6 @@ export default class InfoDisplay extends React.Component<Props> {
 				<div style={typeStyle}>{types.map(type => this.typeColor(type))}</div>
 
 				<p>Bio: {this.props.pokemon.pokemonBio}</p>
-				
-				<p>Moves:</p>
-
-				<div>{moves.map(movesName => {
-						
-						return (<ul style={moveUlStyle}>
-									<li style ={nameListStyle}>{movesName}
-										<li style ={flavorTextStyle}>
-											{createFlavorText(movesIndex)}
-										</li>
-									</li>
-								</ul>
-				)})}
-				</div>
 			</div>
 		);
 	}
@@ -194,15 +165,3 @@ const fairy: React.CSSProperties = {
 	color: "#d685ad"
 };
 
-const moveUlStyle: React.CSSProperties = {
-	listStyleType: "none",
-}
-
-const nameListStyle: React.CSSProperties = {
-	fontWeight: "bold",
-	textTransform: "uppercase",
-}
-const flavorTextStyle: React.CSSProperties = {
-	fontWeight: "normal",
-	textTransform: "none",
-}
